# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "G3 Network API",
    "description": "API nền tảng vận hành xe tải điện — Phase 1 (simulator, dữ liệu giả 100%).\n\nĐăng nhập: POST /auth/otp/request → mã in ra console → POST /auth/otp/verify → gắn header `Authorization: Bearer <token>`.\n\nPhân quyền theo docs/prd/09-rbac.md, mặc định TỪ CHỐI. Truy cập vị trí xe luôn ghi audit log (NF-06, Nghị định 13/2023).",
    "version": "0.2.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/auth/otp/request": {
      "post": {
        "summary": "Xin mã OTP đăng nhập (F-F1)",
        "tags": [
          "xac-thuc"
        ],
        "description": "Luôn trả 202 dù SĐT có tồn tại hay không — người ngoài không dò được danh sách tài khoản. Phase 1: mã in ra console của apps/api, không gửi SMS thật.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone"
                ],
                "properties": {
                  "phone": {
                    "minLength": 8,
                    "maxLength": 20,
                    "description": "SĐT đăng nhập. Phase 1 dùng số GIẢ, xem npm run db:seed (vd 0900000003).",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "description": "Mã lỗi máy đọc được, vd khong_du_quyen",
                          "type": "string"
                        },
                        "message": {
                          "description": "Mô tả tiếng Việt cho người dùng",
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/otp/verify": {
      "post": {
        "summary": "Đổi mã OTP lấy token đăng nhập (F-F1)",
        "tags": [
          "xac-thuc"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "phone",
                  "code"
                ],
                "properties": {
                  "phone": {
                    "minLength": 8,
                    "maxLength": 20,
                    "description": "SĐT đăng nhập. Phase 1 dùng số GIẢ, xem npm run db:seed (vd 0900000003).",
                    "type": "string"
                  },
                  "code": {
                    "minLength": 4,
                    "maxLength": 10,
                    "description": "Mã OTP 6 chữ số",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "access_token",
                    "token_type",
                    "expires_in",
                    "user"
                  ],
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string",
                      "enum": [
                        "Bearer"
                      ]
                    },
                    "expires_in": {
                      "description": "Số giây còn hiệu lực",
                      "type": "integer"
                    },
                    "user": {
                      "type": "object",
                      "required": [
                        "id",
                        "full_name",
                        "role"
                      ],
                      "properties": {
                        "id": {
                          "format": "uuid",
                          "type": "string"
                        },
                        "full_name": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "description": "Mã lỗi máy đọc được, vd khong_du_quyen",
                          "type": "string"
                        },
                        "message": {
                          "description": "Mô tả tiếng Việt cho người dùng",
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Thông tin tài khoản đang đăng nhập & danh sách quyền (F-F1)",
        "tags": [
          "xac-thuc"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "full_name",
                    "role",
                    "customer_id",
                    "permissions"
                  ],
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "full_name": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "customer_id": {
                      "anyOf": [
                        {
                          "format": "uuid",
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "permission",
                          "scope",
                          "require_open_ticket"
                        ],
                        "properties": {
                          "permission": {
                            "type": "string"
                          },
                          "scope": {
                            "description": "own | fleet | all",
                            "type": "string"
                          },
                          "require_open_ticket": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "description": "Mã lỗi máy đọc được, vd khong_du_quyen",
                          "type": "string"
                        },
                        "message": {
                          "description": "Mô tả tiếng Việt cho người dùng",
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}